install: remove --no-rootfs arg (ondev_no_rootfs)
Remove the ability to generate a rootfs and the installer image around
it with one command, and make the behavior of --no-rootfs the default.

The reasoning is, that this makes the (already quite complex) installer
code even more complex, and that this feature isn't used. Installer
images are generated by bpo, which has been using --no-rootfs for a long
time.