install: can't combine --ondev with --split
Add a check to ensure --ondev is not combined with --split. It doesn't
work with the on-device installer (1 and 2), because it assumes that the
boot partition is on the same "storage device" as the installation data:
  [boot partition] [free space] [installation data]
(I'll try to support another partition for cgpt_kpart and ondev2 soon,
it will also be on the same storage device.)

With --split, the boot partition is not on the same "storage device". To
make it work with --split, we would rougly need something like the
following:
* have a separate boot image
* have another image with partition header that contains the following
  partitions:
  [free space] [installation data]
* postmarketos-ondev would need to be able to find where the other image is
  flashed, and be able to use the installation data from there, offer to
  install to the free space there etc.

This is not supported by the ondev1 implementation, and out of scope for
the initial version of ondev2.
AFAIK the only device outside of testing this affects is asus-me176c.