install: add --ondev2-rootfs
Add an option to generate images compatible with ondev2. The big
difference is that the installer now lives in the initramfs-extra in the
boot partition, and so the entire root partition is replaced with the OS
image to be installed. Instead of having the OS image inside an
installer rootfs. This saves space.