pmb.install.partition: add high level function
Rename the old partition() function to partition_regular(), and add a
new partition() function that decides between partition_regular() and
partition_cgpt(). Use that in install_system_image.

This makes it a bit easier to read and leads to less code duplication in
upcoming new function install_sytem_image_ondev2().