install: get_partition_layout: use args.deviceinfo
Let get_partition_layout() figure out if we need an extra kernel
partition from deviceinfo or not on its own, without explicitly passing
it to the function. This makes calling it a bit nicer, and I'll add a
new function that calls it in a follow-up patch.

Usually I don't like to make more code use the big args variable. But
here it makes sense to access deviceinfo directly, and it happens to
still be in args for now. To get rid of args here, in the future we
could put deviceinfo in something like pmb.config.deviceinfo, initialize
it once and use it later on throughout the code.