flasher: QoL improvements
These patches improve the quality of life of pmbootstrap flasher, adding
a flash_all command to flash all relevant partitions, and the ability to
specify additional fastboot commands.

In tandem, these patches enable a new user with a fastboot device to do
something like

pmbootstrap flasher flash_all erase dtbo reboot

to install a locally built postmarketOS image, erase the dtbo partition,
and reboot the device.

fastboot supports many commands in a single invocation, a future
optimisation could allow for the above flasher invocation to "build" a
single fastboot command.

My usecase for this is primarily to flash a new image and reboot
the device in a single command.



# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

Signed-off-by: Caleb Connolly <kc@postmarketos.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lists.sr.ht/~postmarketos/pmbootstrap-devel/%3C20230429-flash_all_extra-v1-0-cbc79ed9e329@postmarketos.org%3E

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20230429-flash_all_extra-9431457ed741",
    "base-branch": "master",
    "prefixes": [],
    "history": {
      "v1": [
        "20230429-flash_all_extra-v1-0-cbc79ed9e329@postmarketos.org"
      ]
    }
  }
}