Make Suffix an explicit type
These patches attempt to kickstart the process of creating a proper API
for pmbootstrap. The chroot "suffix" string follows a well-known format
which is obeyed throughout the codebase, it is made up of a suffix
"type", one of:

* rootfs
* buildroot
* installer
* native

Followed optionally by an underscore ("_") and then a well-known name. The
rules on when a name is required and what values it can have depend on
the suffix type:

* rootfs, installer: name is required and must be a device codename
* buildroot: name is required and must be a valid architecture
* native: name must not be set

This patch series creates a Suffix type to represent this abstraction
and strictly enforce it at runtime.

A future patch will introduce strict suffix validation.

This initial step should enable future work to more strictly define other
common parts of the codebase.

I've been running this patch series locally for a few weeks and updating
it when I've run into issues. I think most of the problems are ironed
out

This series also includes a bunch of fixes for running tests locally.

Depends on https://lists.sr.ht/~postmarketos/pmbootstrap-devel/patches/42611
# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

---
Changes in v3:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v2: https://lists.sr.ht/~postmarketos/pmbootstrap-devel/%3C20230629-suffix-type-hint-v2-0-a02c38e269f2@postmarketos.org%3E

Changes in v2:
- rebase on build/bootstrap patch to fix conflict
- Link to v1: https://lists.sr.ht/~postmarketos/pmbootstrap-devel/%3C20230629-suffix-type-hint-v1-0-e92802ae20a8@postmarketos.org%3E

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20230629-suffix-type-hint-d4428d9c25dd",
    "base-branch": "master",
    "prefixes": [],
    "history": {
      "v1": [
        "20230629-suffix-type-hint-v1-0-e92802ae20a8@postmarketos.org"
      ],
      "v2": [
        "20230629-suffix-type-hint-v2-0-a02c38e269f2@postmarketos.org"
      ]
    }
  }
}