HACK: pmb.chroot: support usrmerge
Because we bootstrap a chroot directly with apk, instead of fetching a
preliminary rootfs, it's non-trivial to init with usrmerge. Instead we
need to do some kinda janky shell-scripting to convert the minimal
rootfs to a merged-usr setup. Add the necessary jank to do this.

Additionally, pass in the local binary repository when running
apk.static so that custom-built versions of alpine-base or dependencies
will be used.

Signed-off-by: Caleb Connolly <kc@postmarketos.org>